pyproject.toml
[tool.poetry]
name = "rag-dropbox"
version = "0.1.0"
description = ""
authors = ["Ema Ilic <ema.ilic9@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
openai = "^1.65.4"
pypdf2 = "^3.0.1"
tqdm = "^4.67.1"
sentence-transformers = "^3.4.1"
python-dotenv = "^1.0.1"
pinecone = "^6.0.1"
pdfplumber = "^0.11.5"
pillow = "^11.1.0"
pytesseract = "^0.3.13"
pdf2image = "^1.17.0"
python-pptx = "^1.0.2"
python-docx = "^1.1.2"
dropbox = "^12.0.2"
fastapi = {extras = ["standard"], version = "^0.115.12"}
uvicorn = "^0.34.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"